Übung 6: CRUD für Spalten, Validation für Spalten und Tasks

+++ app/Controllers/TasksErstellen.php
@@ -31,7 +31,7 @@ class TasksErstellen extends BaseController
         // Damit die Erstellen View weiß, von wo sie herkam.
         $data['table'] = "Table";
 
-        // Abfangen von URL-Manipulationen: leitet zurück zu der Tabellenansicht weiter, wenn kein gültiger Task beim Erstellen/Löschen ausgewählt ist.
+        // Abfangen von URL-Manipulationen: leitet zurück zu der Tabellenansicht weiter, wenn kein gültiger Task beim Bearbeiten/Löschen ausgewählt ist.
         // $data['selected_task'] ist nicht gesetzt, wenn die $taskId ungültig ist.
         // Update und Delete müssen eine *valide* TaskId besitzen, create darf überhaupt keine TaskId besitzen.
         if ((($todo == "update" || $todo == "delete") && !isset($data['selected_task']))  ||
@@ -51,6 +51,13 @@ class TasksErstellen extends BaseController
             $data['todo'] = ($todo == "update") ? "update" : "delete";
         } else{ // Abfangen von URL-Manipulationen: leitet zurück zu der Tabellenansicht weiter, wenn Aktion keine der obigen drei ist.
             return redirect()->to(base_url('public/tasks/table'));
+        }
+
+        if (session('data') !== null) {
+            // Bei gescheiterter Validierung werden hier alle bis dahin korrekt ausgefüllten Werte mit übergeben,
+            // damit diese beim Neuladen der Seite in den Formularfeldern bleiben. Plus die neuen Fehlermeldungen.
+            // session('data') hat bei gleichen Array-Schlüsseln Vorrang vor $data.
+            $data = array_merge($data, session('data'));
         }
 
         echo view('templates/header');
@@ -82,7 +89,7 @@ class TasksErstellen extends BaseController
         // Damit die Erstellen View weiß, von wo sie herkam.
         $data['cards'] = "Cards";
 
-        // Abfangen von URL-Manipulationen: leitet zurück zu dem ersten verfügbaren Board auf der Cards-Ansicht weiter, wenn kein gültiger Task/Board beim Erstellen/Löschen ausgewählt ist.
+        // Abfangen von URL-Manipulationen: leitet zurück zu dem ersten verfügbaren Board auf der Cards-Ansicht weiter, wenn kein gültiger Task/Board beim Erstellen/Bearbeiten/Löschen ausgewählt ist.
         // $data['selected_task'] und $data['selected_board'] sind nicht gesetzt, wenn die entsprechenden $taskId oder $boardId ungültig sind.
         if (!isset($data['selected_board'])){
             return redirect()->to(base_url('public/tasks/cards/' . $data['boards'][0]['id']));
@@ -104,6 +111,13 @@ class TasksErstellen extends BaseController
             $data['todo'] = ($todo == "update") ? "update" : "delete";
         } else{ // Abfangen von URL-Manipulationen: leitet zurück zu dem ersten verfügbaren Board auf der Cards-Ansicht weiter, wenn Aktion keine der obigen drei ist.
             return redirect()->to(base_url('public/tasks/cards/' . $data['selected_board']['id']));
+        }
+
+        if (session('data') !== null) {
+            // Bei gescheiterter Validierung werden hier alle bis dahin korrekt ausgefüllten Werte mit übergeben,
+            // damit diese beim Neuladen der Seite in den Formularfeldern bleiben. Plus die neuen Fehlermeldungen.
+            // session('data') hat bei gleichen Array-Schlüsseln Vorrang vor $data.
+            $data = array_merge($data, session('data'));
         }
 
         echo view('templates/header');
@@ -133,25 +147,68 @@ class TasksErstellen extends BaseController
 
         $tasksModel = new TasksModel();
         $session = session();
-        // Differenzierung zwischen Cards- und Table-Ansicht geschieht in der View
-        if ($todo == "create") {
-            // Die Daten werden nun als neuer Task im Model in die Datenbank eingefügt.
-            $tasksModel->createTask($data);
-            $session->setFlashdata('success', 'Task erstellt!');
-        }
-        elseif ($todo == "update") {
-            // Die Daten werden nun an der Stelle id == $taskId die vorherige Zeile überschreiben, in der Datenbank, durch das Model
-            $tasksModel->updateTask($data, $taskId);
-            $session->setFlashdata('success', 'Task aktualisiert.');
-        }
-        elseif ($todo == "delete") {
-            // Die Datenbank wird nun nach $taskid durchsucht, und dessen Zeile gelöscht.
-            $tasksModel->deleteTask($taskId);
-            // Kein Fehler, Löschen soll nur rot angezeigt werden.
-            $session->setFlashdata('error', 'Task gelöscht.');
-        }
+        // Bei Create und Update wird die Validierung der Eingaben durchgeführt.
+        if ($this->validation->run($_POST,'taskErstellenRules') || $todo == "delete") {
+            // Differenzierung zwischen Cards- und Table-Ansicht geschieht in der View
+            if ($todo == "create") {
+                // Die Daten werden nun als neuer Task im Model in die Datenbank eingefügt.
+                $tasksModel->createTask($data);
+                // Für die Bestätigungsmeldung auf der Tasks-Übersichtsseite, ob Cards oder Table.
+                $session->setFlashdata('success', 'Task erstellt!');
+            }
+            elseif ($todo == "update") {
+                // Die Daten werden nun an der Stelle id == $taskId die vorherige Zeile überschreiben, in der Datenbank, durch das Model
+                $tasksModel->updateTask($data, $taskId);
+                // Für die Bestätigungsmeldung auf der Tasks-Übersichtsseite, ob Cards oder Table.
+                $session->setFlashdata('success', 'Task aktualisiert.');
+            }
+            elseif ($todo == "delete") {
+                // Die Datenbank wird nun nach $taskid durchsucht, und dessen Zeile gelöscht.
+                $tasksModel->deleteTask($taskId);
+                // Für die Bestätigungsmeldung auf der Tasks-Übersichtsseite, ob Cards oder Table.
+                // Kein Fehler, Löschen soll nur rot angezeigt werden.
+                $session->setFlashdata('error', 'Task gelöscht.');
+            }
 
-        // Leitet zurück dorthin, von wo die Erstellen View aufgerufen wurde.
-        return redirect()->to(base_url('public/tasks/' . $view . (($view == "cards") ? '/' . $boardId : '' ) ));
+            // Leitet zurück dorthin, von wo die Erstellen View aufgerufen wurde.
+            return redirect()->to(base_url('public/tasks/' . $view . (($view == "cards") ? '/' . $boardId : '' ) ));
+        }
+        else{
+            // Bei fehlerhafter Validierung werden die bisherigen Eingaben zurück in $data gepackt und an die Erstellen View übergeben,
+            // zum Anzeigen der bis dahin eingegebenen Werten.
+            $data = $_POST;
+            // Beinhaltet die Fehlermeldungen, die in der Erstellen-View rot unter den Eingabefeldern angezeigt werden.
+            $data['error'] = $this->validation->getErrors();
+
+            // Alle relevanten, nicht in den Formularfeldern vorhandenen Daten werden ebenfalls zurückgegeben.
+            $data['todo'] = $todo;
+            // Alle Boards geladen für den Fallback aufs erste verfügbare Board bei URL-Manipulationen.
+            // Zusätzlich wird der gerade ausgewählte Board gespeichert, fürs korrekte Zurückleiten nach Abschließen der Aktion in der Erstellen View
+            $boardsModel = new BoardsModel();
+            $data['boards'] = $boardsModel->getData();
+            $data['selected_board'] = $boardsModel->getDataFromBoard($boardId);
+            // Es werden immer alle Spalten für den Spalten Dropdown geladen.
+            $spaltenModel = new SpaltenModel();
+            $data['spalten'] = $spaltenModel->getData();
+            // Fürs Bearbeiten und Löschen wird der betroffene Task geladen.
+            $tasksModel = new TasksModel();
+            $data['selected_task'] = $tasksModel->getDataFromTask($taskId);
+            // Es werden immer alle Taskarten für den Taskarten Dropdown geladen.
+            $taskartenModel = new TaskartenModel();
+            $data['taskarten'] = $taskartenModel->getData();
+            // Es werden immer alle Personen für den Personen Dropdown geladen.
+            $personenModel = new PersonenModel();
+            $data['personen'] = $personenModel->getData();
+            // Damit die Erstellen View weiß, von wo sie herkam.
+            $data['view'] = $view;
+
+            // Leitet zurück zu genau derselben Erstellen View weiter, wo wir uns gerade befinden, mit den bisherigen Eingaben und den neuen Fehlermeldungen.
+            // Dient dazu, die Seite neu zu laden, damit die Fehlermeldungen angezeigt werden können.
+            // with() erlaubt das Übergeben von Daten in einem redirect, ähnlich zu echo view('pages/spalten-erstellen', $data);
+            return redirect()
+                ->to(base_url('public/tasks-erstellen/' . $view . (($view == "cards") ? '/' . $boardId : '' ) . '/' . $todo . (($todo == 'update' || $todo == 'delete') ? '/' . $taskId: '') ))
+                ->withInput()
+                ->with('data', $data);
+        }
     }
 }
